feat: responsive sidebar done di sales
