+++ components/WelcomingMessage.tsx
@@ -3,10 +3,18 @@
 import { FC, useEffect, useState } from "react";
 import { useAppSelector } from "@/redux/store";
 import Image from "next/image";
+import Sidebar from "@/components/Sidebar";
+
+import { Button } from "@nextui-org/react";
 
 const WelcomingMessage: FC = () => {
   const activeUser = useAppSelector((state) => state.auth.value.username);
   const [user, setUser] = useState("");
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+
+  const toggleSidebar = () => {
+    setIsSidebarOpen(!isSidebarOpen);
+  };
 
   useEffect(() => {
     switch (activeUser) {
@@ -32,13 +40,42 @@ const WelcomingMessage: FC = () => {
   }, [activeUser]);
 
   return (
-    <div className="relative flex items-center justify-between h-[11.5vh] w-full bg-white px-10 py-5 text-black">
+    <div className="relative flex h-[11.5vh] w-full items-center justify-between bg-white px-2 py-5 text-black">
+      {/* Sidebar */}
+      <div
+        className={`fixed left-0 top-0 z-50 h-full w-full transition-transform duration-300 ease-in-out ${
+          isSidebarOpen ? "translate-x-0" : "-translate-x-full"
+        } w-[60vw] md:w-[35.3vw] md:translate-x-0`}
+      >
+        <Sidebar />
+      </div>
+
+      {/* Button to toggle Sidebar */}
+      <Button
+        className={`z-[100] mr-2 h-[40px] w-[40px] min-w-0 rounded-xl px-0 transition-all ease-in-out md:hidden ${isSidebarOpen ? "bg-white" : "bg-[#1e3a8a]"}`}
+        onClick={toggleSidebar}
+      >
+        <svg
+          width="800px"
+          height="800px"
+          viewBox="0 0 20 20"
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          className={`transition-all ${isSidebarOpen ? "rotate-90 transform" : ""}`}
+        >
+          <path
+            fill={isSidebarOpen ? "#2563EB" : "#fff"}
+            fill-rule="evenodd"
+            d="M19 4a1 1 0 01-1 1H2a1 1 0 010-2h16a1 1 0 011 1zm0 6a1 1 0 01-1 1H2a1 1 0 110-2h16a1 1 0 011 1zm-1 7a1 1 0 100-2H2a1 1 0 100 2h16z"
+          />
+        </svg>
+      </Button>
       <div className="flex-grow border-l-2 border-black pl-4">
         <p className="text-l font-semibold">Selamat Datang</p>
         <p className="text-m">{user}</p>
       </div>
 
-      <div className="absolute right-20 top-0">
+      <div className="absolute right-10 top-0 md:right-20">
         <Image
           height={200}
           width={200}
